Fixed Resource Leakage
Any Resource Leakage due to a Window's Form being open and not being closed is now resolved. Completely. Follow this format in future:
this.Hide();
newForm.Show();
this.Close();
this.Dispose();
Or alternatively, if a Form is going back to a form, and the form is merely hidden - try this (may return null if non-existent): Application.OpenForms["FormName"];